refactor: config 파일 관리 패키지로 이동
