+++ src/main/kotlin/notbe/tmtm/ddanddanserver/config/SwaggerConfig.kt
@@ -1,4 +1,4 @@
-package notbe.tmtm.ddanddanserver.presentation.config
+package notbe.tmtm.ddanddanserver.config
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition
 import io.swagger.v3.oas.annotations.servers.Server
